# improve string parsing performance