--- a/src/main/java/org/json/JSONTokener.java
+++ b/src/main/java/org/json/JSONTokener.java
@@ -72,7 +72,7 @@ public class JSONTokener {
      * @param s     A source string.
      */
     public JSONTokener(String s) {
-        this(new StringReader(s));
+        this(new NonSynchronizedStringReader(s));
     }