@stleary 2024-07-13
Closed due to using a modified version of a GPLv2-licensed file (or possibly a modified Oracle-proprietary file, I could not tell from the comments). In general, this project does not allow use of modified licensed files.